util: Refactor quilt-tester to use util.WaitFor
This generalizes the waitFor() function from quilt-tester/exec
to accept an arbitrary timeout and migrates it to the util package
so that other packages can make use of it.
