+++ quilt-tester/exec.go
@@ -15,10 +15,8 @@ import (
 	"github.com/NetSys/quilt/api"
 	"github.com/NetSys/quilt/api/client"
 	"github.com/NetSys/quilt/db"
+	"github.com/NetSys/quilt/util"
 )
-
-// Stored in a variable so we can mock it out for unit tests.
-var sleep = time.Sleep
 
 // runSpecUntilConnected runs the given spec, and blocks until either all
 // machines have connected back to the daemon, or 500 seconds have passed.
@@ -44,7 +42,8 @@ func runSpecUntilConnected(spec string) (string, string, error) {
 		return true
 	}
 
-	return stdout, stderr, waitFor(allMachinesConnected, 8*time.Minute)
+	err = util.WaitFor(allMachinesConnected, 1*time.Second, 8*time.Minute)
+	return stdout, stderr, err
 }
 
 // stop stops the given namespace, and blocks until there are no more machines
@@ -68,7 +67,7 @@ func stop(namespace string) (string, string, error) {
 		return len(instances) == 0
 	}
 
-	return stdout, stderr, waitFor(stopped, 2*time.Minute)
+	return stdout, stderr, util.WaitFor(stopped, 1*time.Second, 2*time.Minute)
 }
 
 // downloadSpecs gets the given import path.
@@ -88,22 +87,6 @@ func runSpec(spec string) (string, string, error) {
 func runQuiltDaemon() {
 	cmd := exec.Command("quilt", "-l", "debug", "daemon")
 	execCmd(cmd, "QUILT")
-}
-
-// waitFor waits until `pred` is satisfied, or `timeout` Duration has passed.
-func waitFor(pred func() bool, timeout time.Duration) error {
-	timeoutChan := time.After(timeout)
-	for {
-		select {
-		case <-timeoutChan:
-			return errors.New("timed out")
-		default:
-			if pred() {
-				return nil
-			}
-		}
-		sleep(1 * time.Second)
-	}
 }
 
 func logAndUpdate(sc bufio.Scanner, l fileLogger, logFmt string) chan string {
